FEAT: Completed lightweight Clipboard test suite
