+++ client/components/ClipboardButton.tsx
@@ -29,7 +29,6 @@ export const ClipboardButton = () => {
         end={ <DoneAllIcon /> }
         onClick={ handleClick }
         testId="bttn-copy"
-        data-testId='bttn-copy'
       />
     </Box>
   );
